+++ src/components/icon/icon.ts
@@ -1,6 +1,6 @@
 import { LitElement, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
-import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
+import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 import { getAttributes, toString } from '@carbon/icon-helpers';
 import IconScss from './icon.scss';
 
@@ -14,31 +14,35 @@ export class Icon extends LitElement {
 
   /** The imported Carbon icon. */
   @property({ type: Object })
-  icon: any = {};
+  accessor icon: any = {};
 
   /** Icon fill color. */
   @property({ type: String })
-  fill = 'currentColor';
+  accessor fill = 'currentColor';
 
   /** Specify a size in pixels to override the imported Carbon icon's original size. */
   @property({ type: Number })
-  sizeOverride = null;
+  accessor sizeOverride!: number;
 
   override render() {
-    const attributes = JSON.parse(JSON.stringify(this.icon.attrs));
-    attributes.fill = this.fill;
+    if (Object.keys(this.icon).length > 0) {
+      const attributes = JSON.parse(JSON.stringify(this.icon.attrs));
+      attributes.fill = this.fill;
 
-    if (this.sizeOverride) {
-      attributes.width = this.sizeOverride;
-      attributes.height = this.sizeOverride;
+      if (this.sizeOverride) {
+        attributes.width = this.sizeOverride;
+        attributes.height = this.sizeOverride;
+      }
+
+      const iconString = toString({
+        ...this.icon,
+        attrs: getAttributes(attributes),
+      });
+
+      return html` ${unsafeHTML(iconString)} `;
+    } else {
+      return null;
     }
-
-    const iconString = toString({
-      ...this.icon,
-      attrs: getAttributes(attributes),
-    });
-
-    return html` ${unsafeHTML(iconString)} `;
   }
 }